#first clean testing scripts completed. using files in a helpers folder adjacent to /tests/functional/. A couple notes: I've made a modification to /lib/db/user.py in the generateActivationHash function: if 'paste.testing_variables' in request.environ: request.environ['paste.testing_variables']['hash_and_email'] = '%s__%s'%(hash, toEmail)
This makes available to webtest the hash and toEmail variables. These are needed to activate a user with the web app after the user is 'signed up'. Also, changes in templates/lib/admin_helpers/member.mako have to do with quoting type and name fields in inputs, which is necessary for webtest to recognize and use them. Finally, I put a try/catch around the redirect in the systemAdmin index function because session['redirect_to'] was failing.